// src/types.ts
export type KeyName =
  | "enter"
  | "tab"
  | "backspace"
  | "delete"
  | "escape"
  | "space"
  | "up"
  | "down"
  | "left"
  | "right"
  | "home"
  | "end"
  | "pageUp"
  | "pageDown"
  | "insert"
  | `f${1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12}`;